docs/example/requirements.txt
django-pages-cms>=0.0.15
django-embed-video==0.11
